skylake: read out and report full width of gen_pmcon registers
GEN_PMCON_A and GEN_PMCON_B are 32-bits wide. Read out and
save the full 32 bits for completeness.

BUG=chrome-os-partner:42847
BRANCH=None
TEST=Built and booted. Noted output on terminal.

Original-Commit-Id: 5a419b2538dc45b1bd0d19b7e6afd45fff9dd4a0
Original-Change-Id: Ie587e886ea34e36d106ff4670781467266a51ddb
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/286270
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: http://review.coreboot.org/11006
Tested-by: build bot (Jenkins)

+++ src/soc/intel/skylake/include/soc/pm.h
@@ -124,8 +124,8 @@ struct chipset_power_state {
 	uint16_t tco2_sts;
 	uint32_t gpe0_sts[4];
 	uint32_t gpe0_en[4];
-	uint16_t gen_pmcon_a;
-	uint16_t gen_pmcon_b;
+	uint32_t gen_pmcon_a;
+	uint32_t gen_pmcon_b;
 	int prev_sleep_state;
 	uint16_t hsio_version;
 	uint16_t hsio_checksum;
